requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
chardet==3.0.4
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
idna==2.8
joblib==0.14.0
numpy==1.17.3
python-dateutil==2.8.2
pytz==2022.4
scipy==1.8.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.25.11
